requirements.txt
torch
pydub
webvtt-py
more-itertools
ffmpeg-python==0.2.0
whisper @ git+https://github.com/openai/whisper.git#egg=whisper